Added a door utility function FindDoorByKeyFlag()
Updated some of the dialog on the sewer computer

The sewer trap can now be completed by the player. They have to use the computer to unlock the first door, then walk to the middle of the second chamber. At that point both doors will close and lock. The way out is to shoot the wall pannel they player passed on the way in.